backend/requirements.txt
Django~=5.0
djangorestframework~=3.15
psycopg2-binary~=2.9
gunicorn~=22.0 # WSGI HTTP Server for Python
django-filter~=24.1
boto3~=1.34 # AWS S3 SDK
django-storages~=1.13 # Django storage backend for S3
whitenoise~=6.6 # For serving static files in production (Heroku)
dj-database-url~=2.1 # For parsing DATABASE_URL
django-storages==1.14.2 # New
boto3==1.34.128 # New
Pillow==10.3.0 # Ensure Pillow is present for ImageField
django-cors-headers==4.3.1
